s2-package {s2}	R Documentation

s2: Spherical Geometry Operators Using the S2 Geometry Library

Description

Provides R bindings for Google's s2 library for geometric calculations on the sphere. High-performance constructors and exporters provide high compatibility with existing spatial packages, transformers construct new geometries from existing geometries, predicates provide a means to select geometries based on spatial relationships, and accessors extract information about geometries.

Author(s)

Maintainer: Edzer Pebesma edzer.pebesma@uni-muenster.de (ORCID)

Authors:

Dewey Dunnington dewey@fishandwhistle.net (ORCID)

Ege Rubak rubak@math.aau.dk

Other contributors:

Jeroen Ooms jeroen.ooms@stat.ucla.edu (configure script) [contributor]

Google, Inc. (Original s2geometry.io source code) [copyright holder]

See Also

Useful links:

https://r-spatial.github.io/s2/

https://github.com/r-spatial/s2

http://s2geometry.io/

Report bugs at https://github.com/r-spatial/s2/issues

[Package s2 version 1.1.6 Index]