S2 Geography Transformations

S2 Geography Transformations

s2_boundary(x)

s2_centroid(x)

s2_closest_point(x, y)

s2_nearest_feature(x, y)

s2_difference(x, y, model = -1L, snap_level = -1L)

s2_sym_difference(x, y, model = -1L, snap_level = -1L)

s2_intersection(x, y, model = -1L, snap_level = -1L)

s2_union(x, y = NULL, model = -1L, snap_level = -1L)

s2_snap_to_grid(x)

s2_union_agg(x, na.rm = FALSE)

s2_centroid_agg(x, na.rm = FALSE)

Arguments

x
An object that can be interpreted as an S2 type.

y
An object that can be interpreted as an S2 type.

model
integer; specifies boundary model; see s2_contains

snap_level
integer; if positive, specifies the snap level; see s2_set_snaplevel

na.rm
For aggregate calculations use na.rm = TRUE to drop missing values.

See also