compute S2Cap or S2LatLngRect of each geography

compute S2Cap or S2LatLngRect of each geography

s2_cap(x)

s2_lat_lng_rect(x)

Arguments

x
object of class s2_geography

Value

Both functions return a data.frame with parameters; s2_cap returns for each geography a bounding spherical cap that contains the region. For points, polylines and geometry collections the bound may not be tight.

s2_lat_lng_rect returns a bounding latitude-longitude rectangle that contains the region. The bound may not be tight for points, polylines and geometry collections. The rectangle returned may depend on the order of points or polylines. lng_lo values larger than lng_hi indicate regions that span the antimeridian, see the Fiji example.

Examples

s2_cap(s2_data_countries("Antarctica"))

#>   lat lng    angle
#> 1 -90   0 26.72934

s2_cap(s2_data_countries("Netherlands"))

#>        lat      lng    angle
#> 1 52.15706 5.203512 1.792879

s2_cap(s2_data_countries("Fiji"))

#>         lat      lng    angle
#> 1 -17.15444 178.7459 1.801369

s2_lat_lng_rect(s2_data_countries("Antarctica"))

#>   lat_lo    lat_hi lng_lo lng_hi   lat_cnt lng_cnt
#> 1    -90 -63.27066   -180    180 -76.63533       0

s2_lat_lng_rect(s2_data_countries("Netherlands"))

#>     lat_lo  lat_hi   lng_lo   lng_hi  lat_cnt  lng_cnt
#> 1 50.80372 53.5104 3.314971 7.092053 52.15706 5.203512

s2_lat_lng_rect(s2_data_countries("Fiji"))

#>      lat_lo    lat_hi  lng_lo    lng_hi   lat_cnt  lng_cnt
#> 1 -18.28799 -16.02088 177.285 -179.7933 -17.15444 178.7459