Create an S2 Geography Vector

Geography vectors are arrays of points, lines, polygons, and/or collections of these. Geography vectors assume coordinates are longitude and latitude on a perfect sphere.

as_s2_geography(x, ...)

s2_geography()

# S3 method for s2_geography
as_s2_geography(x, ...)

# S3 method for s2_lnglat
as_s2_geography(x, ...)

# S3 method for s2_point
as_s2_geography(x, ...)

# S3 method for wk_wkb
as_s2_geography(x, ..., oriented = FALSE, check = TRUE)

# S3 method for WKB
as_s2_geography(x, ..., oriented = FALSE, check = TRUE)

# S3 method for blob
as_s2_geography(x, ..., oriented = FALSE, check = TRUE)

# S3 method for wk_wkt
as_s2_geography(x, ..., oriented = FALSE, check = TRUE)

# S3 method for character
as_s2_geography(x, ..., oriented = FALSE, check = TRUE)

# S3 method for logical
as_s2_geography(x, ...)

# S3 method for s2_geography
as_wkb(x, ...)

# S3 method for s2_geography
as_wkt(x, ...)

Arguments

x
An object that can be converted to an s2_geography vector

...
Unused

oriented
TRUE if polygon ring directions are known to be correct (i.e., exterior rings are defined counter clockwise and interior rings are defined clockwise).

check
Use check = FALSE to skip error on invalid geometries

Value

An object with class s2_geography

Details

The coercion function as_s2_geography() is used to wrap the input of most functions in the s2 package so that you can use other objects with an unambiguious interpretation as a geography vector. Geography vectors have a minimal vctrs implementation, so you can use these objects in tibble, dplyr, and other packages that use the vctrs framework.

See also