s2: Spherical Geometry Operators Using the S2 Geometry Library

Provides R bindings for Google's s2 library for geometric calculations on the sphere. High-performance constructors and exporters provide high compatibility with existing spatial packages, transformers construct new geometries from existing geometries, predicates provide a means to select geometries based on spatial relationships, and accessors extract information about geometries.

See also

Author

Maintainer: Edzer Pebesma edzer.pebesma@uni-muenster.de (ORCID)

Authors:

Dewey Dunnington dewey@fishandwhistle.net (ORCID)

Ege Rubak rubak@math.aau.dk

Other contributors:

Jeroen Ooms jeroen.ooms@stat.ucla.edu (configure script) [contributor]

Google, Inc. (Original s2geometry.io source code) [copyright holder]