S2 Geography Predicates

These functions operate two geography vectors (pairwise), and return a logical vector.

s2_contains(x, y, options = s2_options(model = 0))

s2_within(x, y, options = s2_options(model = 0))

s2_covered_by(x, y, options = s2_options(model = 2))

s2_covers(x, y, options = s2_options(model = 2))

s2_disjoint(x, y, options = s2_options())

s2_intersects(x, y, options = s2_options())

s2_equals(x, y, options = s2_options())

s2_intersects_box(
  x,
  lng1,
  lat1,
  lng2,
  lat2,
  detail = 1000,
  options = s2_options()
)

s2_touches(x, y, options = s2_options())

s2_dwithin(x, y, distance, radius = s2_earth_radius_meters())

Arguments

x
An object that can be interpreted as an S2 type.

y
An object that can be interpreted as an S2 type.

options
An s2_options() object describing the polygon/polyline model to use and the snap level.

lng1, lat1, lng2, lat2
A latitude/longitude range

detail
The number of points with which to approximate non-geodesic edges.

distance
A distance on the surface of the earth in the same units as radius.

radius
Radius of the earth. Defaults to the average radius of the earth as defined by s2_earth_radius_meters().

Model

The geometry model indicates whether a geometry includes its boundaries. Boundaries of line geometries are its end points. OPEN geometries do not contain their boundary (model = 0); CLOSED geometries (model = 2) contain their boundary; HALF-CLOSED geometries contain, like, half of their boundaries, such that when two polygons do not overlap or two lines do not cross, no point exist that belong to more than one of the geometries. (This latter form, half-closed, is not present n the OpenGIS "simple feature access" (SFA) standard, or DE9-IM on which that is based). A value of -1 does not set the model, leaving the S2 default (HALF-CLOSED). The default values for s2_contains() (0) and covers/covered_by (2) correspond to the SFA standard specification of these operators.

See also

Examples

s2_contains(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)")
)

#> [1]  TRUE FALSE

s2_within(
  c("POINT (5 5)", "POINT (-1 1)"),
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))"
)

#> [1]  TRUE FALSE

s2_covered_by(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)")
)

#> [1] FALSE FALSE

s2_covers(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)")
)

#> [1]  TRUE FALSE

s2_disjoint(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)")
)

#> [1] FALSE  TRUE

s2_intersects(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)")
)

#> [1]  TRUE FALSE

s2_equals(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c(
    "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
    "POLYGON ((10 0, 10 10, 0 10, 0 0, 10 0))",
    "POLYGON ((-1 -1, 10 0, 10 10, 0 10, -1 -1))"
  )
)

#> [1]  TRUE  TRUE FALSE

s2_intersects(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)")
)

#> [1]  TRUE FALSE

s2_intersects_box(
  c("POINT (5 5)", "POINT (-1 1)"),
  0, 0, 10, 10
)

#> [1]  TRUE FALSE

s2_touches(
  "POLYGON ((0 0, 0 1, 1 1, 0 0))",
  c("POINT (0 0)", "POINT (0.5 0.75)", "POINT (0 0.5)")
)

#> [1]  TRUE FALSE FALSE

s2_dwithin(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)"),
  0 # distance in meters
)

#> [1]  TRUE FALSE

s2_dwithin(
  "POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))",
  c("POINT (5 5)", "POINT (-1 1)"),
  1e6 # distance in meters
)

#> [1] TRUE TRUE